+++ src/test/java/ladder/domain/ladder/LIneTest.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain.ladder;
 
-import ladder.domain.ladder.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
