+++ src/test/java/ladder/domain/user/UserTest.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain.user;
 
-import ladder.domain.user.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
