+++ src/test/java/ladder/domain/user/UsersTest.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain.user;
 
-import ladder.domain.user.Users;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
